January Bug Fixes & Improvements

+++ apps/api/src/db/course_chapters.py
@@ -1,5 +1,5 @@
 from typing import Optional
-from sqlalchemy import BigInteger, Column, ForeignKey
+from sqlalchemy import Column, ForeignKey, Integer
 from sqlmodel import Field, SQLModel
 
 
@@ -7,10 +7,10 @@ class CourseChapter(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     order: int
     course_id: int = Field(
-        sa_column=Column(BigInteger, ForeignKey("course.id", ondelete="CASCADE"))
+        sa_column=Column(Integer, ForeignKey("course.id", ondelete="CASCADE"))
     )
     chapter_id: int = Field(
-        sa_column=Column(BigInteger, ForeignKey("chapter.id", ondelete="CASCADE"))
+        sa_column=Column(Integer, ForeignKey("chapter.id", ondelete="CASCADE"))
     )
     org_id: int = Field(default=None, foreign_key="organization.id")
     creation_date: str
